moving function to own file
